compiled wieder

+++ Source/DafnyLanguageServer/Handler/LspStandard/DefinitionTaskHandler.cs
@@ -40,8 +40,8 @@ namespace DafnyLanguageServer.Handler
             {
                 return await Task.Run(() =>
                 {
-                    var line = request.Position.Line + 1;
-                    var col = request.Position.Character + 1;
+                    var line = (int)request.Position.Line + 1;
+                    var col = (int)request.Position.Character + 1;
                     var manager = _workspaceManager.SymbolTableManager;
                     var selectedSymbol = manager.GetSymbolByPosition(line, col);
